KMAC1: limit PWM length to k and k+1. Fix KSM k-mer orientation inconsistency; increase motif OR pseudocount to 0.05*#neg.

+++ src/edu/mit/csail/cgs/utils/stats/StatUtil.java
@@ -1176,7 +1176,7 @@ public class StatUtil {
 	 * @param n sample size
 	 * @return
 	 */
-	public static double odds_ratio(int P, int N, int p, int n, double pseudocount) {
+	public static double odds_ratio(int P, int N, double p, double n, double pseudocount) {
 		double or = (p+pseudocount)/(P-p+pseudocount) / ((n+pseudocount)/(N-n+pseudocount));
 		return or;	
 	}
